Implementation of DP algorithm
This algorithm first orders the entry weights, the puts diamonds with same weight inside buckets, and then starts emptying buckets on the ground, from the heaviest to the lightest, and combining them with the next heaviest diamond possible. This is done until only one diamond is left, which means no more combinations can be done. This leftover diamond is the optimal solution seeked.